ТЗ 11-1

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/LikeStorage.java
@@ -14,24 +14,16 @@ public class LikeStorage {
 
     private final JdbcTemplate jdbcTemplate;
 
-
     public void add(long filmId, long user_id) {
-        jdbcTemplate.update(""
-                + "INSERT INTO likes (film_id, user_id) "
-                + "VALUES (?, ?)", filmId, user_id);
-
+        jdbcTemplate.update("INSERT INTO likes (film_id, user_id) VALUES (?, ?)", filmId, user_id);
     }
 
+    //'^[a-z]([a-zA-Z0-9]*)?$'.
     public void delete(long filmId, long user_id) {
-        jdbcTemplate.update(""
-                + "DELETE FROM likes "
-                + "WHERE film_id=? "
-                + "AND user_id=?", filmId, user_id);
-
+        jdbcTemplate.update("DELETE FROM likes WHERE film_id=? AND user_id=?", filmId, user_id);
     }
 
     public Integer count(long filmId) {
         return jdbcTemplate.queryForObject("SELECT COUNT(*) FROM likes WHERE film_id=" + filmId, Integer.class);
     }
-
 }
